+++ src/node/selector.rs
@@ -103,7 +103,7 @@ impl selectors::Element for NodeRef<'_> {
         //TODO: maybe we should unpack compare_node directly here
         self.tree
             .compare_node(&self.id, &other.id, |a, b| {
-                if let (NodeData::Element(ref e1), NodeData::Element(ref e2)) = (&a.data, &b.data) {
+                if let (NodeData::Element(e1), NodeData::Element(e2)) = (&a.data, &b.data) {
                     e1.name == e2.name
                 } else {
                     false